buffer rx y tx para la flash implementados

+++ Principal/flashLib.h
@@ -13,7 +13,7 @@
   void escritura_flash(uint32_t sector_inicial,uint32_t n_sectores, uint16_t numero_palabras_escritura , uint32_t inicio_direccion_escritura , uint32_t *datos);
 	void escritura_flash_string_format(uint32_t sector_inicial,uint32_t n_sectores, uint32_t inicio_direccion_escritura , char *datos);
 	uint32_t* lectura_flash(uint32_t inicio_direccion_escritura, uint16_t palabras_lectura);
-	char* lectura_flash_string_format(uint32_t inicio_direccion_escritura, uint16_t palabras_lectura);
+	char* lectura_flash_string_format(uint32_t inicio_direccion_escritura, int palabras_lectura);
 	int num_palabras_string(char * cadena_caracteres);
 	char* conversion_string(uint32_t *datos, char *str);
